@@ -10,21 +10,28 @@
     "license": "MIT",
     "require": {
         "php": "^8.2",
-        "laravel/framework": "^11.31",
+        "barryvdh/laravel-dompdf": "^3.1",
+        "blade-ui-kit/blade-heroicons": "^2.6",
+        "davidhsianturi/blade-bootstrap-icons": "^2.1",
+        "laravel/framework": "12.*",
+        "laravel/telescope": "^5.11",
         "laravel/tinker": "^2.9",
-        "livewire/livewire": "^3.4",
-        "livewire/volt": "^1.0"
+        "livewire/flux": "^2.4",
+        "livewire/livewire": "^3.6",
+        "livewire/volt": "^1.7",
+        "robsontenorio/mary": "^2.4"
     },
     "require-dev": {
         "fakerphp/faker": "^1.23",
         "fly-apps/dockerfile-laravel": "^1.0",
+        "laravel/boost": "*",
         "laravel/breeze": "^2.2",
         "laravel/pail": "^1.1",
         "laravel/pint": "^1.13",
         "laravel/sail": "^1.26",
         "mockery/mockery": "^1.6",
         "nunomaduro/collision": "^8.1",
-        "pestphp/pest": "^3.6",
+        "pestphp/pest": "3.*",
         "pestphp/pest-plugin-laravel": "^3.0"
     },
     "autoload": {
