@@ -51,7 +51,7 @@ $deleteUser = function (Logout $logout) {
                 <x-input-label for="password" value="{{ __('Password') }}" class="sr-only" />
 
                 <x-text-input
-                    wire:model="password"
+                    wire:model.live="password"
                     id="password"
                     name="password"
                     type="password"
